Add encounters view

+++ src/main/java/ee/a1nu/discord_dkp_bot/database/model/Weight.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @Setter
 public class Weight extends BaseEntity {
     @Column(nullable = false, unique = true, name = "encounter_weight")
-    Integer weight;
+    byte weight;
 }
